git-svn-id: https://svn.code.sf.net/p/liar/code/trunk/liar@96 b153b6fe-5711-0410-bcb9-8978480c73f3

+++ src/output/display.h
@@ -85,7 +85,7 @@ private:
 	typedef std::vector<XYZ> TRenderBuffer;
 	typedef std::vector<TScalar> TWeightBuffer;
 	typedef std::vector<PixelToaster::FloatingPointPixel> TDisplayBuffer;
-	typedef prim::Aabb2D<unsigned> TDirtyBox;
+	typedef prim::Aabb2D<size_t> TDirtyBox;
 
 	enum ToneMapping
 	{
